+++ src/commands/arg_structs/info_args.rs
@@ -3,8 +3,10 @@ use crate::aliases::ModID;
 
 #[derive(Args, Debug, Clone)]
 pub struct ModInfoArgs {
+   
     
-    pub mod_id: ModID,
+    #[arg(num_args = 1..)]
+    pub mod_id: Vec<ModID>,
    
     /// Shows the description of the mod, Note: This can take a lot of space in the terminal
     #[arg(short = 'd', long)]